+++ test/fixtures/channel-message-filter.ts
@@ -8,17 +8,6 @@ import { channel, privateChat, supergroup } from "./chats";
 import { adminUser, bot, systemChannelBot } from "./users";
 
 /**
- * Webhook response which contains answer callback query method.
- * It should be sent as a result of callback save settings processing.
- */
-export const answerCbSaveSettingsWebhookResponse = {
-  callback_query_id: "1",
-  method: "answerCallbackQuery",
-  show_alert: true,
-  text: "Changes saved",
-};
-
-/**
  * Payload for ban sender chat request. It should be sent as a result of channel message
  * processing in a supergroup chat.
  */