+++ test/fixtures/settings.ts
@@ -70,6 +70,17 @@ export const addedToNewChatSendMessagePayload2 = {
 };
 
 /**
+ * Webhook response which contains answer callback query method.
+ * It should be sent as a result of callback save settings processing.
+ */
+export const answerCbSaveSettingsWebhookResponse = {
+  callback_query_id: "1",
+  method: "answerCallbackQuery",
+  show_alert: true,
+  text: "Changes saved",
+};
+
+/**
  * Payload for edit message text request. It should be sent as a result of chats callback.
  */
 export const cbChatsEditMessageTextPayload = {