+++ test/warnings.e2e-spec.ts
@@ -7,6 +7,7 @@ import { server } from "test/utils/server";
 
 import { AppModule } from "../src/app.module";
 import { privateChat, supergroup } from "./fixtures/chats";
+import * as settingsFixtures from "./fixtures/settings";
 import { adminUser, bot, user } from "./fixtures/users";
 import * as fixtures from "./fixtures/warnings";
 import {
@@ -163,7 +164,7 @@ describe("WarningsModule (e2e)", () => {
     const response = await request(TEST_WEBHOOK_BASE_URL).post(TEST_WEBHOOK_PATH).send(fixtures.cbSaveSettingsWebhook);
 
     expect(response.status).toBe(200);
-    expect(response.body).toEqual(fixtures.answerCbSaveSettingsWebhookResponse);
+    expect(response.body).toEqual(settingsFixtures.answerCbSaveSettingsWebhookResponse);
     await sleep(ASYNC_REQUEST_DELAY);
     expect(editMessageTextPayload).toEqual(fixtures.cbSaveSettingsEditMessageTextPayloadFunc());
   });