+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.17.0 (2024-xx-xx)
 
 - Fix: duplicate message about completion of voteban
+- Fix: incorrect display of the vote limit in the voteban settings
 - Other: internal improvements
 
 ## 1.16.0 (2024-02-12)